+++ Workspace/PandemicStats/src/main/java/br/edu/unoesc/pandemicstats/springboot/repository/EmpresaRepository.java
@@ -6,11 +6,11 @@ import org.springframework.stereotype.Repository;
 import br.edu.unoesc.pandemicstats.springboot.model.Empresa;
 
 @Repository
-public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
 	
 	@Query(value = "select e from Empresa e where e.cnpjemp = ?1")
-	Empresa findByCNPJ(int cnpj);
+	Empresa findByCNPJ(long cnpj);
 	
 	@Query(value = "select e from Empresa e where e.cpfusu.cpfusu = ?1")
-	Empresa findByCpfusu(int cpfusu);
+	Empresa findByCpfusu(long cpfusu);
 }
