+++ Workspace/PandemicStats/src/main/java/br/edu/unoesc/pandemicstats/springboot/repository/PacienteRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import br.edu.unoesc.pandemicstats.springboot.model.Paciente;
 
 @Repository
-public interface PacienteRepository extends JpaRepository<Paciente, Integer> {
+public interface PacienteRepository extends JpaRepository<Paciente, Long> {
 	
 	@Query(value = "select p from Paciente p where p.cpfusu.cpfusu = ?1")
-	Paciente findByCPF(int cpfusu);
+	Paciente findByCPF(long cpfusu);
 }
