+++ Workspace/PandemicStats/src/main/java/br/edu/unoesc/pandemicstats/springboot/schemmas/ReqEmpSCH.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class ReqEmpSCH {
-	private int cnpjemp;
+	private long cnpjemp;
 }
