+++ Workspace/PandemicStats/src/main/java/br/edu/unoesc/pandemicstats/springboot/schemmas/ShowCidSCH.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class ShowCidSCH {
 	private String nomcid;
-	private int codest;
+	private long codest;
 	
 	public void Convert(Cidade c)
 	{
